streamlit/requirements.txt
streamlit
boto3
scikit-learn==1.6.1
joblib
Pillow
